requirements.txt
Flask==2.0.1
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1
numpy==1.21.2